before setting up authentication

+++ components/elements/inspector/element.tsx
@@ -7,8 +7,9 @@ import Image from "next/image"
 import style from "./style.module.scss"
 import { useAppDispatch, deleteFromMeta } from "../../../state-store"
 
+import { motion } from 'framer-motion'
 import Editor from "./editor"
-import { Remove, Delete } from "../buttons"
+import { Delete } from "../buttons"
 
 
 const Element = ({ data, index }: prop) => {
@@ -19,21 +20,43 @@ const Element = ({ data, index }: prop) => {
 
     const onClick = () => { setExpended(!isExpanded) }
 
-    const onDelete = () => {
+    const onDelete = (e: any) => {
 
-        dispatch(deleteFromMeta(index))
+        e.stopPropagation()
+
+        const permission = confirm('do you really want to detele them')
+
+        permission && dispatch(deleteFromMeta(index));
+
     }
 
-    const imageSrcConditonally = isExpanded ? '/chevron-up.svg' : "/chevron-down.svg"
+    const imageSrcConditonally = isExpanded ? '/chevron-up.svg' : "/chevron-down.svg";
+
+    const isNameError = (data.name == '') ? true : false;
+
 
     return (
-        <div className={style.element}>
+        <div
+            style={{ background: isNameError ? '#ff000017' : 'white' }}
+            className={style.element}>
 
-            <div onClick={onClick} className={style.header}>
-                <span>{index + 1}. {data.name}  <p>{data.action}</p></span>
+            <div
+                onClick={onClick}
+                className={style.header}>
+
+                <span >
+                    {index + 1}. {data.name}
+                    <p>{data.action}</p></span>
                 <div>
-                    <Delete aspect={30} color="red" onClick={onDelete} />
-                    <Image src={imageSrcConditonally} width={15} height={15} alt="noen" />
+                    <Delete
+                        aspect={30}
+                        color="red"
+                        onClick={onDelete} />
+                    <Image
+                        src={imageSrcConditonally}
+                        width={15}
+                        height={15}
+                        alt="noen" />
                 </div>
             </div>
